dispense with JSInternalFrame
